requirements.txt
asgiref==3.8.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
Markdown==3.7
pillow==10.4.0

solana
requests
jwt
cryptography
